refactor: split candidate_compression.rs into focused modules (#939)
Split the monolithic candidate_compression.rs (1,136 lines) into a
candidate_compression/ directory with focused sub-modules:

- mod.rs: public API, CompressibleGroup, UUID generation, re-exports
- grouping.rs: candidate grouping and deduplication
- identity.rs: IDENTITY neuron compression (Issue #921)
- nonlinear.rs: TANH/GELU non-linear compression (Issue #922)
- gain_estimation.rs: saturation-aware gain estimation

Public API unchanged — all existing use paths work via re-exports.
All 25 unit tests and 19 integration tests pass without modification.
